pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "linshare-mcp"
version = "0.1.0"
description = "LinShare MCP Server"
requires-python = ">=3.10"
dependencies = [
    "mcp",
    "fastmcp",
    "requests",
    "python-dotenv",
    "httpx",
    "uvicorn",
    "chainlit"
]

[project.scripts]
# Entry points
